refactor files & new id system for reports
