+++ prisma/schema.prisma
@@ -56,7 +56,7 @@ model Paragraph {
 }
 
 model Report {
-  id      String       @id @default(cuid())
+  number  Int
   user    User         @relation(name: "Report_User", fields: [userId], references: [id])
   userId  String
   guildId String
@@ -75,6 +75,7 @@ model Report {
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
 
-  @@index([userId, guildId])
+  @@id(name: "id", [number, guildId])
+  @@index([userId, guildId, number])
   @@map("reports")
 }
