+++ src/bot.ts
@@ -3,30 +3,42 @@ import { PrismaClient } from '@prisma/client';
 import { ActivityType, IntentsBitField, Partials } from 'discord.js';
 import { Client } from 'discordx';
 import 'dotenv/config';
-import { dirname, importCommands } from './commandImporter.js';
+import { dirname, importCommands } from './command.importer.js';
 
 export class NPLAYModerationBot {
 	private static _client: Client;
 	private static _prismaClient: PrismaClient;
 
+	/**
+	 * The discord.js client
+	 */
 	static get Client(): Client {
 		return this._client;
 	}
 
+	/**
+	 * The Prisma client
+	 */
 	static get db(): PrismaClient {
 		return this._prismaClient;
 	}
 
+	/**
+	 * Initialize the Prisma client
+	 */
 	static initDB(): void {
 		this._prismaClient = new PrismaClient();
 	}
 
+	/**
+	 * Start the bot
+	 */
 	static async start(): Promise<void> {
 		const guildID = process.env.GUILD_ID;
 
 		if (!guildID) {
 			console.warn(
-				'GUILD_ID not found in environment variables, bot will fall back to global commands'
+				'GUILD_ID not found in environment variables, bot will fall back to global modules'
 			);
 		}
 
@@ -65,7 +77,7 @@ export class NPLAYModerationBot {
 			this._client.executeInteraction(interaction);
 		});
 
-		await importCommands(`${dirname(import.meta.url)}/commands/**/*.{js,ts}`);
+		await importCommands(`${dirname(import.meta.url)}/modules/**/*.{js,ts}`);
 
 		if (!process.env.BOT_TOKEN) {
 			throw Error('Could not find BOT_TOKEN in your environment');
@@ -79,12 +91,18 @@ export class NPLAYModerationBot {
 		this._prismaClient.$disconnect();
 	}
 
+	/**
+	 * Restart the bot
+	 */
 	static async restart(): Promise<void> {
 		console.log('Restarting bot...');
 		this.destroy();
 		process.exit(1);
 	}
 
+	/**
+	 * Stop the bot
+	 */
 	static async stop(): Promise<void> {
 		console.log('Stopping bot...');
 		this.destroy();
