+++ src/modules/permission/permission.guard.ts
@@ -1,9 +1,9 @@
 import { CommandInteraction, GuildMember, PermissionsBitField } from 'discord.js';
 import { GuardFunction } from 'discordx';
-import { CommandOnlyOnGuildError } from '../embed/data/genericEmbeds.js';
-import { NoPermissionEmbed } from '../embed/data/permissionEmbeds.js';
-import { createEmbed } from '../embed/embed.js';
-import { getUserPermissions, hasPermission } from './permissionHelpers.js';
+import { CommandOnlyOnGuildError } from '../../embed/data/genericEmbeds.js';
+import { NoPermissionEmbed } from '../../embed/data/permissionEmbeds.js';
+import { createEmbed } from '../../embed/embed.js';
+import { getUserPermissions, hasPermission } from './permission.service.js';
 import { Permission, PermissionBitmapFlags } from './permissions.js';
 
 /**
